% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{combine_matrices}
\alias{combine_matrices}
\title{Combine a list of matrices}
\usage{
combine_matrices(matrices, as_edge_list = FALSE)
}
\arguments{
\item{matrices}{list of matrices}

\item{as_edge_list}{If \code{TRUE}, return a table with columns for the row,
column, and value for every unique cell in the combined matrix instead of a
matrix.}
}
\description{
Combine a list of matrices
}
\keyword{internal}
