% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{considered_ref_meta_parsed}
\alias{considered_ref_meta_parsed}
\title{Find and parse NCBI reference metadata}
\usage{
considered_ref_meta_parsed(path = NULL, family = NULL, json_path = NULL)
}
\arguments{
\item{path}{The path to one or more folders that contain pathogensurveillance
output.}

\item{family}{The names of families to return data for.}

\item{json_path}{Path to one or more JSON files containing reference metadata
to parse. It is assumed that these files are named by families. This is an
alternative input to the \code{path} parameter and both can not be used at once.}
}
\value{
A table of reference metadata
}
\description{
Returns the parsed NCBI reference metadata for all references considered for
download.
}
