% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_finders.R
\name{estimated_ani_matrix_path}
\alias{estimated_ani_matrix_path}
\title{Find the estimated ANI matrix file path}
\usage{
estimated_ani_matrix_path(path, simplify = FALSE, format = "csv")
}
\arguments{
\item{path}{The path to one or more folders that contain pathogensurveillance
output.}

\item{simplify}{If \code{TRUE}, return a single vector with all paths rather than
a list of vectors for each output directory.}

\item{format}{One of \code{csv}, \code{npy}, or \code{npy.labels.txt}. Formats other than
\code{csv} are only available when using whole output directories as input.}
}
\value{
character vector of length 1
}
\description{
Return the file path to the CSV with the approximate ANI (average nucleotide
identity) matrix estimated by sourmash for a given pathogensurveillance
output folder.
}
\examples{
path <- system.file('extdata/ps_output', package = 'PathoSurveilR')
estimated_ani_matrix_path(path)

}
\seealso{
Other path finders: 
\code{\link{busco_ref_path}()},
\code{\link{busco_tree_path}()},
\code{\link{considered_ref_meta_path}()},
\code{\link{core_ref_path}()},
\code{\link{core_tree_path}()},
\code{\link{downloaded_ref_meta_path}()},
\code{\link{pocp_matrix_path}()},
\code{\link{ref_meta_path}()},
\code{\link{run_info_path}()},
\code{\link{sample_meta_path}()},
\code{\link{sendsketch_path}()},
\code{\link{software_version_path}()},
\code{\link{status_message_path}()},
\code{\link{variant_align_path}()},
\code{\link{variant_ref_path}()},
\code{\link{variant_tree_path}()}
}
\concept{path finders}
