% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_finders.R
\name{find_output_directory_path}
\alias{find_output_directory_path}
\title{Find pipeline output directory paths}
\usage{
find_output_directory_path(path)
}
\description{
For one or more paths, return the paths of subfolders (or the given folders)
that are the entire output of a pathogensurviellance run.
}
\keyword{internal}
