% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_finders.R
\name{find_path}
\alias{find_path}
\title{Search for a specific type of file in pathogensurveillance output}
\usage{
find_path(
  path,
  out_dir_subpath,
  out_dir_pattern,
  report_dir_subpath,
  report_dir_pattern,
  simplify = TRUE
)
}
\arguments{
\item{path}{One or more paths to search for pathogensurveillance output,
either entire output directories or report input directories.}

\item{out_dir_subpath}{A path relative to \code{path} that does not change. Used
to lessen the number of files that need to be searched through.}

\item{out_dir_pattern}{A regex matching the path of the file/directory
relative to \code{path}/\code{out_dir_subpath}.}

\item{report_dir_subpath}{A path relative to \code{path} that does not change.
Used to lessen the number of files that need to be searched through.}

\item{report_dir_pattern}{A regex matching the path of the file/directory
relative to \code{path}/\code{report_dir_subpath}.}

\item{simplify}{If \code{TRUE}, return a single vector with all paths rather than
a list of vectors for each output directory.}
}
\description{
Search for a specific type of file in pathogensurveillance output
}
\keyword{internal}
