% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_finders.R
\name{find_path_data}
\alias{find_path_data}
\title{Find paths for specific pathogensurveillance output files}
\usage{
find_path_data(outdir_path, targets = NULL, simplify = TRUE, long = TRUE)
}
\arguments{
\item{outdir_path}{The path to an output directory of pathogensurviellance}

\item{targets}{The names of one or more output types to search for.}

\item{simplify}{If \code{TRUE}, combine all of the output found into a single
table using the \code{long} option to decide how to format it.}

\item{long}{If \code{TRUE} when combining output into a single table, put all
paths in a single \code{path} column rather than having a column for each
\code{target}.}
}
\description{
Find paths for specific pathogensurveillance output files and associated
metadata
}
