% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_finders.R
\name{find_path_in_output_directory}
\alias{find_path_in_output_directory}
\title{Find files in pathogensurveillance output}
\usage{
find_path_in_output_directory(path, subpath, pattern)
}
\arguments{
\item{path}{One or more paths to search.}

\item{subpath}{A path relative to \code{path} that does not change. Used to lessen
the number of files that need to be searched through.}

\item{pattern}{A regex matching the path of the file/directory relative to
\code{path}/\code{subpath}.}
}
\description{
Find files in the output of pathogensurveillance.
}
\keyword{internal}
