% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{find_ps_data}
\alias{find_ps_data}
\title{Parse pathogensurveillance output files}
\usage{
find_ps_data(path, target, simplify = FALSE, must_exist = !is.null(target))
}
\arguments{
\item{target}{The names of one or more output types to search for. Setting to
\code{NULL} causes data for all available paths to be returned.}

\item{simplify}{If \code{TRUE}, attempt to combine all of the output found into a
single object instead of returning a list.}

\item{must_exist}{If \code{TRUE}, an error is generated if the specified target
does not exist. Otherwise \code{NULL} is returned for missing \code{target}s}
}
\description{
Parse data for specific pathogensurveillance output files
}
