% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_finders.R
\name{find_ps_paths}
\alias{find_ps_paths}
\title{Find paths for specific pathogensurveillance output files}
\usage{
find_ps_paths(
  path,
  target,
  simplify = TRUE,
  long = TRUE,
  all_metadata = FALSE,
  must_exist = !is.null(target)
)
}
\arguments{
\item{target}{The names of one or more output types to search for. Setting to
\code{NULL} causes data for all available paths to be returned.}

\item{simplify}{If \code{TRUE}, combine all of the output found into a single
table using the \code{long} option to decide how to format it.}

\item{long}{If \code{TRUE}, put all paths in a single column rather than having
each target in their own column.}

\item{all_metadata}{Include columns for all metadata regarding the discovery
and parsing of files.}

\item{must_exist}{If \code{TRUE}, an error is generated if the specified target
does not exist. Otherwise \code{NULL} is returned for missing \code{target}s}

\item{outdir_path}{One or more paths to output directories of specific output
files of pathogensurviellance}
}
\description{
Find paths for specific pathogensurveillance output files and associated
metadata
}
