% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_finders.R
\name{find_static_path}
\alias{find_static_path}
\title{Find files in pathogensurveillance output}
\usage{
find_static_path(path, output_path, report_path, return_contents = TRUE, ...)
}
\arguments{
\item{path}{One or more paths to search}

\item{output_path}{The path of the file/directory relative to the entire
pathogensurveillance output.}

\item{report_path}{The path of the file/directory relative to the entire
report input directory.}

\item{return_contents}{If \code{TRUE} (the default), if \code{path} is a directory, return the files
in that directory rather than the directory itself.}
}
\description{
Find files in the output of pathogensurveillance, either in the entire output
folder or the report input folder.
}
\keyword{internal}
