% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{generalized_tree_plot}
\alias{generalized_tree_plot}
\title{Plot generic phylogeny}
\usage{
generalized_tree_plot(
  path = NULL,
  target = NULL,
  tree = NULL,
  sample_meta = NULL,
  ref_meta = NULL,
  collapse_by_tax = NULL,
  interactive = FALSE
)
}
\arguments{
\item{path}{The path to one or more folders that contain pathogensurveillance
output or paths to tree files.}

\item{target}{The names of one or more output types to search for.}

\item{tree}{One or more trees to plot}

\item{sample_meta}{The path to sample metadata or the data itself in the form
of a table.}

\item{ref_meta}{The path to reference metadata or the data itself in the form
of a table.}

\item{collapse_by_tax}{If not \code{NULL} and more that one \code{tree} is supplied,
use taxonomy data passed to this option to combine the trees together.}

\item{interactive}{Whether to use an HTML-based interactive format or not
(default: TRUE)}
}
\value{
A list of plots, unless \code{collapse_by_tax} is used, in which case a
single plot is returned.
}
\description{
Plot phylogenies present in the output of a pathogensurveillance run.
}
