% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{generic_matrix_parsed}
\alias{generic_matrix_parsed}
\title{Parse distance matrices}
\usage{
generic_matrix_parsed(
  path,
  path_func,
  simplify = TRUE,
  as_edge_list = FALSE,
  ...
)
}
\arguments{
\item{path}{The path to one or more folders that contain pathogensurveillance
output.}

\item{path_func}{A path finder function}

\item{simplify}{If \code{TRUE}, combine output into a single matrix, filling in
the gaps with \code{NA}.}

\item{as_edge_list}{If \code{TRUE}, return a table with columns for the row,
column, and value for every unique cell in the combined matrix instead of a
matrix.}

\item{...}{passed to read.table}
}
\value{
a \code{list} of matrices
}
\description{
Parse distance matrices
}
\keyword{internal}
