% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{get_count}
\alias{get_count}
\title{Parse "count" arguments which can be a number or a percentage}
\usage{
get_count(n_choices, count)
}
\description{
Parse "count" arguments which can be a number or a percentage
}
\keyword{internal}
