% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_finders.R
\name{get_ps_outdir_from_path}
\alias{get_ps_outdir_from_path}
\title{Get parent directory that is a pathogensurveillance output directory}
\usage{
get_ps_outdir_from_path(path)
}
\arguments{
\item{path}{A path to a file/folder within a pathogensurveillance output directory}
}
\description{
Get parent directory that is a pathogensurveillance output directory
}
