% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_finders.R
\name{known_outputs}
\alias{known_outputs}
\title{Get which outputs are available}
\usage{
known_outputs(outdir_path, exists = FALSE)
}
\arguments{
\item{outdir_path}{The path to an output directory of pathogensurviellance}

\item{exists}{If \code{TRUE}, information is only returned for output type that
exist, rather than all output type that are known.}
}
\description{
Get which outputs are available for a given pathogensurveillance output
directory.
}
