% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{lookup_ncbi_taxon_id_classification}
\alias{lookup_ncbi_taxon_id_classification}
\title{Lookup classifications using taxon IDs}
\usage{
lookup_ncbi_taxon_id_classification(
  taxon_ids,
  match_input = TRUE,
  simplify = FALSE
)
}
\arguments{
\item{match_input}{If \code{TRUE}, reduplicate the output to match 1:1 with the input.}

\item{simplify}{If \code{TRUE}, combine list of tibbles into a single tibble}
}
\description{
Lookup classifications using taxon IDs
}
\keyword{internal}
