% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_printers.R
\name{make_best_match_table}
\alias{make_best_match_table}
\title{Print best match table}
\usage{
make_best_match_table(pairwise_matrix, sample_data, ref_data)
}
\arguments{
\item{sample_data}{The sample metadata}

\item{ref_data}{The reference metadata}

\item{pairwise_matrices}{A \code{list} of \code{matrix} with all pairwise comparisons
between samples and references}
}
\description{
Prints a table with the highest value for each sample given a pairwise
distance matrix.
}
\keyword{internal}
