% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_data_finders.R
\name{make_path_data_with_group}
\alias{make_path_data_with_group}
\title{Make a table with paths and report group}
\usage{
make_path_data_with_group(path, path_func, regex, id_types, sep = "_")
}
\arguments{
\item{path_func}{A path finder function that returns a list of paths}

\item{regex}{A regular expression with a single capture group that matches
the portion of the file name that is non-constant.}

\item{id_types}{One or more of \code{'sample_id'}, \code{'reference_id'},
\code{'report_id'} in the order that they appear in file name.}

\item{sep}{The separator that is used between elements in the file name.}
}
\description{
Make a table with paths and report group
}
\keyword{internal}
