% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_ncbi_assembly_metadata}
\alias{parse_ncbi_assembly_metadata}
\title{Find and parse NCBI reference metadata}
\usage{
parse_ncbi_assembly_metadata(path = NULL)
}
\arguments{
\item{path}{The path to one or more folders that contain pathogensurveillance
output.}
}
\value{
A table of reference metadata
}
\description{
Returns the parsed NCBI reference metadata for all references considered for
download.
}
