% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_assemblies.R
\name{pick_assemblies}
\alias{pick_assemblies}
\title{Pick assemblies to provide context for one or more taxa}
\usage{
pick_assemblies(
  path = NULL,
  metadata = NULL,
  taxon = NULL,
  rank = NULL,
  subtaxon_count = 5,
  assembly_count = 1,
  rank_col = c("family", "genus", "species", "organism_name"),
  criteria = c("-is_atypical", "is_type", "is_refseq", "is_binomial", "is_annotated",
    "assembly_level", "checkm_completeness", "-checkm_contamination", "-contig_l50",
    "coverage"),
  allow_ambiguous = TRUE,
  allow_atypical = TRUE,
  simplify = TRUE
)
}
\arguments{
\item{path}{The path to one or more folders that contain pathogensurveillance
output. The needed files will be automatically found. Alternatively, the
\code{metadata} and \code{taxon} parameters can be used.}

\item{metadata}{\code{data.frame}/\code{tibble}: Metadata on assemblies that can be
chosen. This is an alternative to the path \code{path} input.}

\item{taxon}{\code{character} vector: One or more taxa to find assemblies for.
This is an alternative to the \code{path} input.}

\item{rank}{\code{character} vector: The rank corresponding to each value in the
\code{taxon} argument. Generally not needed but can be used to avoid unexpected
behavior from the same taxon name being used in multiple ranks. Must be
composed of values present in \code{rank_col}. If not supplied, taxon names
appearing in multiple ranks will be treated as different taxa.}

\item{subtaxon_count}{\code{integer}/\code{character} vector: The number or percentage
(i.e., a \code{character} ending with \verb{\%}) of taxa to get assemblies for one
rank more specific than the rank of the corresponding value in \code{taxon}.
This parameter can either be one value for each value in \code{taxon} or a
single value for all taxa. For example, if \verb{subtaxon_count = c('3', 100\%')}
and \code{taxon = c('Escherichia', 'Lactobacillaceae')} then assemblies for 3
species within Escherichia and all genera within Lactobacillaceae will be
returned.}

\item{assembly_count}{\code{integer}/\code{character} vector: The number or percentage
(i.e., a \code{character} ending with \verb{\%}) of assemblies returned to represent
each subtaxon. This parameter can either be one value for each value in
\code{taxon} or a single value for all taxa.}

\item{rank_col}{\code{character} vector: The names of columns in \code{metadata}
containing taxon names at each rank, in order from broad to specific (e.g.
\code{c('family', 'genus', 'species')}). Should include one rank more specific
than any of the taxa defined in \code{taxon}.}

\item{criteria}{\code{character} vector or \code{list}: The criteria used to decide
which assemblies are the best. If a \code{character} vector then values are
column names in \code{metadata} in which higher numbers are better and \code{TRUE} is
better than \code{FALSE}. Prefixing the column name with \code{-} will reverse this
making lower numbers better, etc. If a \code{list} is supplied, then each item
should be a vector corresponding to rows in \code{metadata} in which higher
numbers are better and \code{TRUE} is better than \code{FALSE}. In both cases,
elements should be sorted in order of decreasing importance.}

\item{allow_ambiguous}{Allow assemblies with things like 'uncultured',
'unknown', 'sp.', or 'incertea sedis' in taxon names.}

\item{allow_atypical}{Allow assemblies with numbers or unusual characters in
taxon names.}
}
\value{
A copy of \code{metadata} with additional columns \code{selected},
\code{selection_rank}, \code{selection_taxon}, and  \code{selection_subtaxon}.
}
\description{
Pick assemblies to provide context for one or more taxa. For each taxon
supplied, representative
}
