% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{plot_map_widget}
\alias{plot_map_widget}
\title{Plot map widget}
\usage{
plot_map_widget(
  path = NULL,
  width = NULL,
  height = NULL,
  elementID = NULL,
  def_size = "",
  def_color = ""
)
}
\arguments{
\item{path}{The path to one or more folders that contain pathogensurveillance
output.}

\item{width}{The width of the widget.}

\item{height}{The height of the widget.}

\item{elementID}{Widget element ID.}

\item{def_size}{The default marker size-by variable}

\item{def_color}{The default marker color-by variable.}
}
\value{
A map plot.
}
\description{
Plot map widget
}
