% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{pocp_matrix_parsed}
\alias{pocp_matrix_parsed}
\title{Get POCP matrix}
\usage{
pocp_matrix_parsed(path, simplify = FALSE, as_edge_list = FALSE)
}
\arguments{
\item{path}{The path to one or more folders that contain pathogensurveillance
output.}

\item{simplify}{If \code{TRUE}, combine output into a single matrix, filling in
the gaps with \code{NA}.}

\item{as_edge_list}{If \code{TRUE}, return a table with columns for the row,
column, and value for every unique cell in the combined matrix instead of a
matrix.}
}
\value{
a \code{list} of ANI matrices or edge lists depending on the values of \code{simplify} and \code{as_edge_list}
}
\description{
Return a list of \code{\link[base:matrix]{base::matrix()}} with the POCP values based on a core
gene analysis using pirate.
}
\examples{
path <- system.file('extdata/ps_output', package = 'PathoSurveilR')
pocp_matrix_parsed(path)

}
