% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{postprocess_table_list}
\alias{postprocess_table_list}
\title{Prepare list of tables for output}
\usage{
postprocess_table_list(table_list, simplify)
}
\arguments{
\item{table_list}{The list to combine}

\item{simplify}{If \code{FALSE} a list of \code{\link[tibble:tibble]{tibble::tibble()}}s are returned named by the
output folder the data was found in. If \code{TRUE}, all data is combined into a
single \code{\link[tibble:tibble]{tibble::tibble()}}.}
}
\description{
Optionally combine a list of tables into a single table for use with path
data finder functions. Also can add columns for which output folders paths
came from.
}
\keyword{internal}
