% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_finders.R
\name{print_outdir_schema}
\alias{print_outdir_schema}
\title{Print pathogensurveillance outdir schema}
\usage{
print_outdir_schema(outdir_path, exists = TRUE)
}
\arguments{
\item{outdir_path}{The path to an output directory of pathogensurviellance}

\item{exists}{If \code{TRUE}, information is only returned for output type that
exist, rather than all output type that are known.}
}
\description{
Print the output directory layout for a given pipeline output
}
