% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_finders.R
\name{print_ps_outputs}
\alias{print_ps_outputs}
\title{Prints available pipeline output files}
\usage{
print_ps_outputs(outdir_path, exists = TRUE)
}
\arguments{
\item{outdir_path}{The path to an output directory of pathogensurviellance}

\item{exists}{If \code{TRUE}, information is only returned for output type that
exist, rather than all output type that are known.}
}
\value{
Invisibly returns a nested list of categories with their items and descriptions.
Primarily called for its side effect of printing formatted output to the console.
}
\description{
Prints available pipeline output files grouped by type
}
