% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{sample_distribution_map}
\alias{sample_distribution_map}
\title{Make map widget}
\usage{
sample_distribution_map(path)
}
\arguments{
\item{path}{The path to one or more folders that contain
pathogensurveillance output or a sample metadata table with coordinates.}
}
\description{
Converts classifications of top hits in sendsketch output into an interactive
sunburst plot.
}
