% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_printers.R
\name{selected_ref_meta_table}
\alias{selected_ref_meta_table}
\title{Print a table of selected reference metadata}
\usage{
selected_ref_meta_table(input, target, interactive = FALSE, ...)
}
\arguments{
\item{target}{Either 'contextual_refs' or 'mapping_refs'}

\item{interactive}{Whether to use an HTML-based interactive format or not (default: TRUE)}

\item{...}{Passed to \code{DT::datatable}.}

\item{path}{The path to one or more folders that contain
pathogensurveillance output.}
}
\description{
Selects columns in the reference metadata to print and format the result for
use as a static table in PDF or an interactive table in HTML.
}
\examples{
path <- system.file('extdata/ps_output', package = 'PathoSurveilR')
reference_meta_table(path)
reference_meta_table(path, interactive = TRUE)

}
