% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{software_version_parsed}
\alias{software_version_parsed}
\title{Parse Software Version Metadata from YAML File}
\usage{
software_version_parsed(path, simplify = TRUE)
}
\arguments{
\item{path}{The path to one or more folders that contain
pathogensurveillance output.}

\item{simplify}{If \code{FALSE} a list of \code{\link[tibble:tibble]{tibble::tibble()}}s are returned named by the
output folder the data was found in. If \code{TRUE}, all data is combined into a
single \code{\link[tibble:tibble]{tibble::tibble()}}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with columns \code{module}, \code{program}, and \code{version}
detailing the software versions.
}
\description{
Reads a YAML file containing software version information and transforms it
into a tibble. Each software module and program, along with its version, is
extracted and organized.
}
\examples{
path <- system.file('extdata/ps_output', package = 'PathoSurveilR')
software_version_parsed(path)

}
\seealso{
Other parsers: 
\code{\link{sendsketch_parsed}()},
\code{\link{sendsketch_taxa_present}()},
\code{\link{sendsketch_taxonomy_parsed}()},
\code{\link{variant_align_parsed}()},
\code{\link{variant_tree_parsed}()}
}
\concept{parsers}
