% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{sort_df_by_columns}
\alias{sort_df_by_columns}
\title{Sort a table by columns, reversing when prefixed with "-"}
\usage{
sort_df_by_columns(df, cols)
}
\description{
Sort a table by columns, reversing when prefixed with "-"
}
\keyword{internal}
