% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{status_message_parsed_summary}
\alias{status_message_parsed_summary}
\title{Get summary of pipeline status data}
\usage{
status_message_parsed_summary(path, simplify = TRUE)
}
\arguments{
\item{path}{The path to one or more folders that contain
pathogensurveillance output.}

\item{simplify}{If \code{FALSE} a list of \code{\link[tibble:tibble]{tibble::tibble()}}s are returned named by the
output folder the data was found in. If \code{TRUE}, all data is combined into a
single \code{\link[tibble:tibble]{tibble::tibble()}}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with counts of messages attributes if
\code{simplify = TRUE} or a list of such \code{\link[tibble:tibble]{tibble::tibble()}}s for each output
directory found if \code{simplify = FALSE}.
}
\description{
Return a \code{\link[tibble:tibble]{tibble::tibble()}} (table) with the numbers of issues encountered by the
pipeline. The contents of all status message files found in the given paths
will be combined.
}
\examples{
path <- system.file('extdata/ps_output', package = 'PathoSurveilR')
status_message_parsed_summary(path)

}
