% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{suggest_option}
\alias{suggest_option}
\title{Suggest correct option names for potentially misspelled inputs}
\usage{
suggest_option(
  input,
  options,
  max_suggestions = 5,
  max_dist = 3,
  ignore_case = TRUE,
  method = "osa"
)
}
\arguments{
\item{input}{The user's input (potentially misspelled option name)}

\item{options}{A character vector of valid option names}

\item{max_suggestions}{Maximum number of suggestions to return}

\item{max_dist}{Maximum allowed edit distance for suggestions}

\item{ignore_case}{Whether to ignore case when matching}
}
\value{
A character vector of suggested corrections, or NULL if no close matches found
}
\description{
Suggest correct option names for potentially misspelled inputs
}
\keyword{internal}
