% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{variant_align_parsed}
\alias{variant_align_parsed}
\title{Find and parse SNP alignments}
\usage{
variant_align_parsed(path, rename = TRUE, simplify = TRUE)
}
\arguments{
\item{path}{The path to one or more folders that contain pathogensurveillance output.}

\item{rename}{If \code{TRUE}, rename the sequence labels as sample IDs and reference IDs.}

\item{simplify}{If \code{FALSE} a list of \code{\link[tibble:tibble]{tibble::tibble()}}s are returned named by the
output folder the data was found in. If \code{TRUE}, all data is combined into a
single \code{\link[tibble:tibble]{tibble::tibble()}}.}
}
\value{
list of \code{\link[ape:DNAbin]{ape::DNAbin()}}, named by alignment file
}
\description{
Returns parsed SNP alignments for each reference used in the
variant analysis for a given pathogensurveillance output folder.
}
\examples{
path <- system.file('extdata/ps_output', package = 'PathoSurveilR')
variant_align_parsed(path)

}
\seealso{
Other parsers: 
\code{\link{sendsketch_parsed}()},
\code{\link{sendsketch_taxa_present}()},
\code{\link{sendsketch_taxonomy_parsed}()},
\code{\link{variant_tree_parsed}()}
}
\concept{parsers}
