% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_data_finders.R
\name{variant_align_path_data}
\alias{variant_align_path_data}
\title{Find the SNP alignment path data}
\usage{
variant_align_path_data(path, simplify = TRUE)
}
\arguments{
\item{path}{The path to one or more folders that contain
pathogensurveillance output.}

\item{simplify}{If \code{FALSE} a list of \code{\link[tibble:tibble]{tibble::tibble()}}s are returned named by the
output folder the data was found in. If \code{TRUE}, all data is combined into a
single \code{\link[tibble:tibble]{tibble::tibble()}}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with one row per path if \code{simplify = TRUE} or a list of
such \code{\link[tibble:tibble]{tibble::tibble()}}s for each output directory found if \code{simplify = FALSE}.
}
\description{
Return a table with the file paths of the SNP alignments for each reference
used in the variant analysis for a given pathogensurveillance output folder.
}
\examples{
path <- system.file('extdata/ps_output', package = 'PathoSurveilR')
variant_align_path_data(path)

}
\seealso{
Other path tables: 
\code{\link{busco_ref_path_data}()},
\code{\link{busco_tree_path_data}()},
\code{\link{considered_ref_meta_path_data}()},
\code{\link{core_ref_path_data}()},
\code{\link{core_tree_path_data}()},
\code{\link{downloaded_ref_meta_path_data}()},
\code{\link{estimated_ani_matrix_path_data}()},
\code{\link{pocp_matrix_path_data}()},
\code{\link{ref_meta_path_data}()},
\code{\link{run_info_path_data}()},
\code{\link{sample_meta_path_data}()},
\code{\link{sendsketch_path_data}()},
\code{\link{software_version_path_data}()},
\code{\link{status_message_path_data}()},
\code{\link{variant_ref_path_data}()},
\code{\link{variant_tree_path_data}()}
}
\concept{path tables}
