% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{combine_data_frames}
\alias{combine_data_frames}
\title{Combines multiple data frames}
\usage{
combine_data_frames(...)
}
\arguments{
\item{...}{One or more data.frames to combine}
}
\description{
combines multiple data frames, handling cases where they have different columns by filling missing values with NA
}
\keyword{internal}
