% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{considered_ref_meta_parsed}
\alias{considered_ref_meta_parsed}
\title{Find and parse NCBI reference metadata}
\usage{
considered_ref_meta_parsed(path = NULL, taxon_id = NULL, simplify = TRUE)
}
\arguments{
\item{path}{The path to one or more folders that contain pathogensurveillance
output.}

\item{taxon_id}{The NCBI taxon IDs of families to return data for.}

\item{simplify}{If \code{FALSE} a list of \code{\link[tibble:tibble]{tibble::tibble()}}s are returned named by the
output folder the data was found in. If \code{TRUE}, all data is combined into a
single \code{\link[tibble:tibble]{tibble::tibble()}}.}
}
\value{
A table of reference metadata
}
\description{
Returns the parsed NCBI reference metadata for all references considered for
download.
}
