% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{find_ps_data}
\alias{find_ps_data}
\title{Parse pathogensurveillance output files}
\usage{
find_ps_data(path, target, simplify = FALSE)
}
\arguments{
\item{target}{The names of one or more output types to search for.}

\item{simplify}{If \code{TRUE}, attempt to combine all of the output found into a
single object instead of returning a list.}
}
\description{
Parse data for specific pathogensurveillance output files
}
