% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_finders.R
\name{find_ps_paths}
\alias{find_ps_paths}
\title{Find paths for specific pathogensurveillance output files}
\usage{
find_ps_paths(
  path,
  target = NULL,
  simplify = TRUE,
  long = TRUE,
  all_metadata = FALSE
)
}
\arguments{
\item{target}{The names of one or more output types to search for.}

\item{simplify}{If \code{TRUE}, combine all of the output found into a single
table using the \code{long} option to decide how to format it.}

\item{long}{If \code{TRUE}, put all paths in a single column rather than having
each target in their own column.}

\item{all_metadata}{Include columns for all metadata regarding the discovery
and parsing of files.}

\item{outdir_path}{The path to an output directory of pathogensurviellance}
}
\description{
Find paths for specific pathogensurveillance output files and associated
metadata
}
