% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{sendsketch_taxa_present}
\alias{sendsketch_taxa_present}
\title{Get taxa predicted by sendsketch results}
\usage{
sendsketch_taxa_present(path, sample_id = NULL, only_best = FALSE)
}
\arguments{
\item{path}{The path to one or more folders that contain pathogensurveillance
output.}

\item{sample_id}{One or more sample IDs to return data for. By default, data
for all samples is returned.}

\item{only_best}{Only return taxa predicted by the best hit for each
combination of report group and sample.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with one row per taxon found
}
\description{
Return which taxa are predicted to be present given sendsketch results
associated with directory paths containing \code{pathogensurveillance} output.
}
\seealso{
Other parsers: 
\code{\link{busco_tree_parsed}()},
\code{\link{core_tree_parsed}()},
\code{\link{multigene_tree_parsed}()},
\code{\link{sendsketch_parsed}()},
\code{\link{sendsketch_taxonomy_data_parsed}()},
\code{\link{sendsketch_taxonomy_parsed}()},
\code{\link{software_version_parsed}()},
\code{\link{tree_parsed}()},
\code{\link{variant_align_parsed}()},
\code{\link{variant_tree_parsed}()}
}
\concept{parsers}
