% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{sendsketch_taxonomy_parsed}
\alias{sendsketch_taxonomy_parsed}
\title{Get parsed sendsketch taxonomy}
\usage{
sendsketch_taxonomy_parsed(
  input,
  ranks_as_cols = TRUE,
  only_shared = FALSE,
  update_taxonomy = FALSE,
  sample_id = NULL,
  only_best = FALSE,
  simplify = TRUE
)
}
\arguments{
\item{input}{The path to one or more folders that contain pathogensurveillance
output or a table in the form of \code{\link[=sendsketch_parsed]{sendsketch_parsed()}} output.}

\item{ranks_as_cols}{If \code{TRUE}, return a table with a column for each rank
instead of a table with one taxon per row.}

\item{only_shared}{If \code{TRUE}, only return the ranks that are present in all
of the inputs. Only has an effect if \code{as_table} is \code{TRUE}.}

\item{update_taxonomy}{If \code{TRUE} look up the current taxonomy classification
using the NCBI taxon ID rather than using the existing one.}

\item{sample_id}{One or more sample IDs to parse the results for. Default:
include all samples.}

\item{only_best}{Only return the best hit for each combination of report
group and sample.}

\item{simplify}{If \code{FALSE} a list of \code{\link[tibble:tibble]{tibble::tibble()}}s are returned named by the
output folder the data was found in. If \code{TRUE}, all data is combined into a
single \code{\link[tibble:tibble]{tibble::tibble()}}.}

\item{include_sample_id}{description}
}
\value{
A \code{\link[base:character]{base::character()}} vector of taxonomic classifications, each
delimited with \verb{;}, named by sample IDs.
}
\description{
Return the taxonomic classification in sendsketch results associated with
directory paths containing \code{pathogensurveillance} output.
}
\examples{
path <- system.file('extdata/ps_output', package = 'PathoSurveilR')
sendsketch_taxonomy_parsed(path)

}
\seealso{
Other parsers: 
\code{\link{sendsketch_parsed}()},
\code{\link{sendsketch_taxa_present}()}
}
\concept{parsers}
